requirements.txt
arrow==1.3.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
cookiecutter==2.4.0
coverage==7.3.2
exceptiongroup==1.1.3
greenlet==3.0.0
hupper==1.12
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.2
PasteDeploy==3.0.1
pip==23.2.1
plaster-pastedeploy==1.0.1
plaster==1.1.2
pluggy==1.3.0
Pygments==2.16.1
pyramid-debugtoolbar==4.10
pyramid-jinja2==2.10
pyramid-mako==1.1.0
pyramid==2.0.2
pytest-cov==4.1.0
pytest==7.4.2
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
setuptools==68.2.2
six==1.16.0
slashdb==0.0
soupsieve==2.5
SQLAlchemy==2.0.21
text-unidecode==1.3
tomli==2.0.1
transaction==3.1.0
translationstring==1.4
types-python-dateutil==2.8.19.14
typing-extensions==4.8.0
urllib3==2.0.6
venusian==3.0.0
waitress==2.1.2
WebOb==1.8.7
WebTest==3.0.0
wheel==0.41.2
zope.deprecation==5.0
zope.interface==6.0
zope.sqlalchemy==3.1
